sales-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
PyJWT
requests